chain/ethereum: Avoid unnecessary metrics locking
Having a mutex around the metrics requires all parts of the indexing
process to synchronize, which will have a detrimental effect on indexing
performance, mainly when indexing historic blocks, where a lot of things
are done in parallel.

Thanks to metrics using `AtomicF64` internally, none of them need to be
mutable and thus require no locking.